fix : 질병 키워드 정보 리스트 타입으로 변경

See merge request s06-bigdata-dist-sub2/S06P22D108!188

+++ backend/src/main/java/com/helloz/lemonaid/db/repository/DiseaseRepository.java
@@ -16,6 +16,6 @@ public interface DiseaseRepository extends JpaRepository<Disease, Long> {
             " group by sd.disease.no" +
             " having count(sd.disease.no) >= :symptomSize")
 
-    Page<Disease> findAllBySymptoms(List<String> symptoms, int symptomSize, Pageable pageable);
+    Page<Disease> findAllBySymptoms(List<String> symptoms, long symptomSize, Pageable pageable);
 
 }
